requirements.txt
# LangChain Core
langchain

# Cohere Integration
langchain-cohere
cohere

# Vector Store
chromadb
langchain-chroma
langchain_community  # For Chroma vector store integration

# PDF Handling
pypdf

# UI
streamlit

# Environment Variable Management
python-dotenv
